+++ aux_ui/.env
@@ -1,4 +1,4 @@
 CLIENT_ID=316d39c75e9c4e13a93c937f6f6b7477
 REDIRECT_URL=http://localhost:4000/callback
-IP_ADDRESS=10.0.0.64
+IP_ADDRESS=10.0.0.67
 CLIENT_SECRET=2232b1d95a824660a097ef618be40b0a
